+++ src/main/java/com/softeq/crawler/Main.java
@@ -1,5 +1,10 @@
 package com.softeq.crawler;
 
+import com.softeq.crawler.jsoup.JsoupHandler;
+import com.softeq.crawler.service.HtmlWebCrawler;
+
+import java.io.IOException;
+
 public class Main {
 
     // Input Data:
@@ -9,8 +14,13 @@ public class Main {
     private static int countPages = 1000;
 
     public static void main(String[] args) {
-        HtmlWebCrawler crawler = new HtmlWebCrawler(PATTERNS);
-        crawler.setLinkDeap(0);
-        System.out.println(crawler.start(ROOT_URI));
+//        HtmlWebCrawler crawler = new HtmlWebCrawler(PATTERNS);
+//        crawler.setLinkDeap(0);
+//        System.out.println(crawler.start(ROOT_URI));
+        try {
+            new JsoupHandler().parse(ROOT_URI);
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
     }
 }
